% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worklogs.R
\name{as_tibble.worklogs_leaf}
\alias{as_tibble.worklogs_leaf}
\title{Collect a Worklogs Node Into a Tibble}
\usage{
\method{as_tibble}{worklogs_leaf}(x, ...)
}
\arguments{
\item{x}{A \code{worklogs_node}.}

\item{...}{Arguments that are passed on to \code{as_tibble}.}
}
\description{
Extract the worklogs from a \code{worklogs_leaf} and apply \code{as_tibble} to it.
}
