% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worklogs.R
\name{as_tibble.worklogs_node}
\alias{as_tibble.worklogs_node}
\title{Collect a Worklogs Node Into a Tibble}
\usage{
\method{as_tibble}{worklogs_node}(x, ...)
}
\arguments{
\item{x}{A \code{worklogs_node}.}

\item{...}{Arguments that are passed on to \code{as_tibble}.}
}
\description{
Collects all of the worklogs in a worklogs tree into a tibble.
}
