% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-worklogs.R
\docType{class}
\name{classes_worklogs}
\alias{classes_worklogs}
\alias{worklogs_node-class}
\title{Configuration Classes}
\description{
S4 classes used for representing worklogs.
}
\section{Slots}{

\describe{
\item{\code{children}}{A named list of \code{worklogs} objects.}

\item{\code{fold_status}}{Either \code{"unfolded"} or \code{"folded"}.}

\item{\code{prototype}}{A data frame with 0 rows used to specify the form of all
children worklog entries.}
}}

