% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-config.R, R/classes-worklogs.R
\docType{class}
\name{classes_worklogs_config}
\alias{classes_worklogs_config}
\alias{config_labels-class}
\alias{worklogs_config-class}
\alias{worklogs_leaf-class}
\alias{worklogs-class}
\title{Configuration Classes}
\description{
S4 classes used in specifying worklogs configuration information.
}
\section{Slots}{

\describe{
\item{\code{description}}{A string specifying the column name of the worklog
description.}

\item{\code{start}}{A string specifying the column name of the worklog start times.}

\item{\code{end}}{A string specifying the column name of the worklog end times.}

\item{\code{duration}}{A string specifying the column name of the worklog durations.}

\item{\code{tags}}{A string specifying the column name of the worklog tags.}

\item{\code{labels}}{A \code{config_labels} object.}

\item{\code{worklogs}}{A data frame of worklog entries.}

\item{\code{name}}{A string providing the name of the worklogs.}

\item{\code{config}}{A \code{worklogs_config} object.}
}}

