% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worklogs.R
\name{effort_summary}
\alias{effort_summary}
\title{Display a Summary of the Efforts in the Worklogs}
\usage{
effort_summary(wkls, show_all = FALSE, effort_style = "percent")
}
\arguments{
\item{wkls}{A \code{worklogs} object.}

\item{show_all}{Either \code{TRUE} or \code{FALSE} specifying whether or not to display
elements of the worklog tree that didn't have any effort.}

\item{effort_style}{One of either "\code{effort}", \code{"percent"} or
\code{"effort_and_percent"} specifying what information to provide in the effort
summary.}
}
\value{
Returns \code{NULL}.
}
\description{
Display a Summary of the Efforts in the Worklogs
}
