% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worklogs-extract.R
\name{extract_worklogs}
\alias{extract_worklogs}
\alias{extract_worklogs,worklogs-method}
\title{Extract a Subtree From a Worklogs Tree}
\usage{
extract_worklogs(wkls, path)

\S4method{extract_worklogs}{worklogs}(wkls, path)
}
\arguments{
\item{wkls}{A \code{worklogs} object.}

\item{path}{A character vector such that each element in vector corresponds}
}
\description{
Extract a Subtree From a Worklogs Tree
}
