% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worklogs.R
\name{fold}
\alias{fold}
\alias{fold,worklogs-method}
\alias{fold_children}
\alias{fold_children,worklogs-method}
\title{Fold Worklogs Below a Given Location}
\usage{
fold(wkls, path)

\S4method{fold}{worklogs}(wkls, path)

fold_children(wkls, path)

\S4method{fold_children}{worklogs}(wkls, path)
}
\arguments{
\item{wkls}{A \code{worklogs} object.}

\item{path}{A character vector such that each element in vector corresponds
to the name of an element in the \code{worklogs} tree.}
}
\value{
A \code{worklogs} object.
}
\description{
Nodes in a worklogs tree can be flagged as being \emph{folded}, which has the
effect of letting functions that operate on worklogs know to consider all
descendents of that node to belong to that node.
}
