% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worklogs-remove.R
\name{remove_worklogs}
\alias{remove_worklogs}
\alias{remove_worklogs,worklogs_node-method}
\alias{remove_worklogs,worklogs_leaf-method}
\title{Remove a Subtree From a Worklogs Tree}
\usage{
remove_worklogs(wkls, path)

\S4method{remove_worklogs}{worklogs_node}(wkls, path)

\S4method{remove_worklogs}{worklogs_leaf}(wkls, path)
}
\arguments{
\item{wkls}{A \code{worklogs} object.}

\item{path}{A character vector such that each element in vector corresponds}
}
\description{
Remove a Subtree From a Worklogs Tree
}
