% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worklogs.R
\name{show,worklogs-method}
\alias{show,worklogs-method}
\title{Display a \code{worklogs} Object}
\usage{
\S4method{show}{worklogs}(object)
}
\arguments{
\item{object}{A \code{worklogs} object.}
}
\description{
Display a \code{worklogs} Object
}
