% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worklogs.R
\name{worklogs}
\alias{worklogs}
\alias{worklogs,list-method}
\alias{worklogs,data.frame-method}
\title{Create a \code{worklogs} Object}
\usage{
worklogs(wkls, split_dfs, config)

\S4method{worklogs}{list}(wkls, split_dfs, config)

\S4method{worklogs}{data.frame}(wkls, split_dfs, config)
}
\arguments{
\item{wkls}{Either a data frame of worklog entries, or a named list}

\item{split_dfs}{Either \code{TRUE} or \code{FALSE}.}

\item{config}{A \code{worklogs_config} object.}
}
\value{
A \code{worklogs} object.
}
\description{
This is one of several available functions that can be used to create
\code{worklogs} objects.
}
