% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-config.R
\name{worklogs_config}
\alias{worklogs_config}
\title{Create a \code{worklogs_config} Object}
\usage{
worklogs_config(
  description_label,
  start_label = NULL,
  end_label = NULL,
  duration_label = NULL,
  tags_label = NULL
)
}
\arguments{
\item{description_label}{A string specifying the column name of the worklog
descriptions.}

\item{start_label}{Either a string specifying the column name of the worklog
start times, or \code{NA_character_} if there is no such column.}

\item{end_label}{Either a string specifying the column name of the worklog
end times, or \code{NA_character_} if there is no such column.}

\item{duration_label}{Either a string specifying the column name of the
worklog duration entries, or \code{NA_character_} if there is no such column.}

\item{tags_label}{Either a string specifying the column name of the worklog
tags, or \code{NA_character_} if there is no such column.}
}
\value{
A \code{worklogs_config} object.
}
\description{
This is the intended mechanism for creating \code{worklogs_config} objects.
}
